requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.136.0
h11==0.16.0
idna==3.11
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
pydantic==2.13.2
pydantic_core==2.46.2
python-dateutil==2.9.0.post0
python-multipart==0.0.26
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.44.0
sqlalchemy==2.0.25
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0